Migrate native code build environment from vs2005 to vs2015

+++ native/dll/DShowCapture-2/SageDTV.cpp
@@ -1626,7 +1626,7 @@ void SageTunerStop( void* Capture )
 	pCapInfo->pMC->Stop();
 }
 
-void SageDelay( void* Capture, unsigned long ms )
+void SageDelay(void* Capture, uint32_t ms)
 {
 	Sleep( ms );
 }
